Change health to be mostly an arbitary integer (but intended 0 to 100)

+++ heaven/src/Api/ServerResponse/NewState.hs
@@ -21,7 +21,7 @@ data IntermediatePlayer =
   IntermediatePlayer { name :: Text
                      , invincible :: Bool
                      , position :: IntermediatePosition
-                     , health :: Double
+                     , health :: Int
                      }
   deriving (Show)
